+++ src/yt_session_computer_assets.c
@@ -75,8 +75,8 @@ yt_session_computer_avoid(struct yt_session *session, struct yt_error *error)
 		return false;
 	if (route != YT_COMPUTER_AVOID_SELECTION_ACCEPTED)
 		return true;
-	if (!yt_computer_avoid_maximum(session_port_offset(session),
-	    session_sector_offset(session), &maximum, error))
+	if (!yt_computer_avoid_maximum((float)session_port_offset(session),
+	    (float)session_sector_offset(session), &maximum, error))
 		return false;
 	{
 		char maximum_text[64];
