+++ src/yt_session_computer_route.c
@@ -88,8 +88,8 @@ yt_session_computer_route(struct yt_session *session, bool autopilot,
 		return true;
 	if (!yt_computer_path_parse(response, &destination_value, error))
 		return false;
-	if (!yt_computer_path_maximum(session_port_offset(session),
-	    session_sector_offset(session), &maximum, error))
+	if (!yt_computer_path_maximum((float)session_port_offset(session),
+	    (float)session_sector_offset(session), &maximum, error))
 		return false;
 	if (destination_value < 1.0f || destination_value > maximum
 	    || start_value < 1.0f || start_value > maximum) {
